# Optimal packing xlaqr5