--- a/SRC/claqr5.f
+++ b/SRC/claqr5.f
@@ -225,7 +225,7 @@
 *> \author Univ. of Colorado Denver
 *> \author NAG Ltd.
 *
-*> \date Januari 2021
+*> \date January 2021
 *
 *> \ingroup complexOTHERauxiliary
 *