--- a/SRC/dlaqr0.f
+++ b/SRC/dlaqr0.f
@@ -390,7 +390,7 @@
 *
 *        ==== Optimal workspace = MAX(DLAQR5, DLAQR3) ====
 *
-         LWKOPT = MAX( NSR / 2, INT( WORK( 1 ) ) )
+         LWKOPT = MAX( 3*NSR / 2, INT( WORK( 1 ) ) )
 *
 *        ==== Quick return in case of workspace query. ====
 *