closes #476 

This PR improves xlaqr5 as described in  "Optimally Packed Chains of Bulges in Multishift QR Algorithms" by L. Karlsson et al.

It is nearly identical to the code included in the paper, but i have reimplemented it to make sure all the indices are correct and verified whether they are the same as the version in the paper.

In the paper, it is mentioned that a few very exceptional cases are still not correctly handled.
These cases appear to be `TST1 = 0` or `TST2 = 0`. In those cases H( K, K-1), H( K, K-2), ...
are accessed, but these elements have not necessarily been computed yet. However,
this is a very exceptional case, and there is to my knowledge no real mathematical advantage
of one technique over the other.

In the original implementation by Byers, great care is taken to assure that the updates within the computational
window from te left are (to the extent possible) column first access. This makes the loop more complicated
and seems to make compiler optimizations harder. I cannot understate how important this is. Without reversing
these loops, this PR would barely make the code faster.

As discussed in the issue. The optimal packing technique makes exploiting the structure of U harder. I have removed support
for that option since it led to small speedups anyway.

The major differences between my implementation and that in the paper are as follows:
- Updates comments
- I've included a small improvement to the accumulation of the reflections in U. Since U is initialized to a unit matrix in every iteration, we can save a few operations by only multiplying the nonzero elements.
- A small change in line 596 to avoid array out of bounds errors. Does not actually affect the execution of the algorithm.

Checklist for things that still need attention:

- [x] Complex implementations
- [x] Check implications on `NTINY`
- [x] Double check vigilant deflation strategy
- [x] Run more tests
- [x] Update comments about workspace requirements in `XHSEQR`

I also want to take to the apportunity to express just how amazing it think the current QR implementation is. I hope that if he's up there, prof. Byers knows I look up to him.

For now, this is just a draft PR; opened because i want to use the CI and the diff. Not ready for review.