@thijssteel 2021-02-01
I did some timings, results look pretty good.
This is linked with intel MKL.

# Old code
randomly generated hessenberg matrix

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      1.7160773      1.1398622      0.5758439        4.055E-15
    1414      2.8858964      1.6754731      1.2096093        4.083E-15
    2000      4.9362059      2.7163010      2.2179706        3.846E-15
    2828      9.6914320      3.6377581      6.0493699        4.023E-15
    4000      7.8365497      7.8237394      0.0000000        2.687E-15
    5657     11.5958040     11.5579986      0.0000000        2.415E-15
    8000     26.4589110     26.4040680      0.0000000        2.386E-15

Hessenberg matrix with A(i,j) = i + j

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      2.8951947      1.1688104      1.7258738        5.996E-15
    1414      5.4274944      1.8647036      3.5618272        6.374E-15
    2000     10.4313257      2.6935471      7.7355328        6.148E-15
    2828     20.5895752      3.4277217     17.1563773        6.205E-15
    4000     53.1000079      8.4607070     44.6227080        6.762E-15
    5657    112.7451473     10.8321977    101.8626941        7.012E-15
    8000    270.5989750     27.0142277    243.5108180        7.393E-15

Laplacian matrix (diag(-1,2,-1))

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      2.3337352      0.7865287      1.5467844        4.192E-14
    1414      4.4392314      1.1337490      3.3046589        4.163E-14
    2000      8.9867607      1.6969975      7.2878094        4.490E-14
    2828     18.4829344      2.3377988     16.1400704        4.404E-14
    4000     48.7693183      5.5059433     43.2498634        4.875E-14
    5657    108.4927636      8.6384931     99.8113961        5.752E-14
    8000    267.4601785     19.0828630    248.3023642        1.159E-13

# New code
randomly generated hessenberg matrix

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      1.3370391      1.0108357      0.3258079        4.028E-15
    1414      2.0243841      1.4623237      0.5612378        3.632E-15
    2000      3.6799751      2.2068826      1.4712231        3.856E-15
    2828      6.0791831      3.1718111      2.9030586        3.720E-15
    4000      6.5873858      6.5739031      0.0000000        2.485E-15
    5657      9.7974927      9.7581081      0.0000000        2.346E-15
    8000     23.4393708     23.3866027      0.0000000        2.365E-15

Hessenberg matrix with A(i,j) = i + j

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      1.9257348      0.9095690      1.0156599        5.472E-15
    1414      3.4931278      1.4147907      2.0772891        5.573E-15
    2000      6.5527296      1.9737693      4.5767572        5.667E-15
    2828     12.8550920      2.9285844      9.9208574        5.822E-15
    4000     34.5621649      6.2815685     28.2646792        6.409E-15
    5657     74.3505805     10.1725267     64.1308556        6.610E-15
    8000    181.5282407     20.3361695    161.1186342        7.276E-15

Laplacian matrix (diag(-1,2,-1))

       n        time(s)    aed time(s)  sweep time(s)            error
    1000      1.5647748      0.6571640      0.9072336        3.924E-14
    1414      2.8630611      0.9965619      1.8656713        4.231E-14
    2000      5.6536871      1.4666950      4.1842327        4.376E-14
    2828     11.2331512      1.9911213      9.2374217        4.424E-14
    4000     31.8281060      5.0413100     26.7723453        5.248E-14
    5657     71.2790017      7.1316567     64.1059008        5.610E-14
    8000    180.5356675     18.4125278    162.0636776        1.265E-13