@langou 2021-02-02
Hi @thijssteel, 

This looks great to me. This is amazing work.

Two questions:

1) Does NTINY in chseqr, claqr0 and claqr4 have to be the same? I see you changed it from 11 to 13. My

2) In \date Januari 2021, can you spell January with a ``y``?

I am very excited that the merging of the pull request is close. At this point, it would be good to have some extra pair of eyes and some people testing this some more. I will try to get more testing in for these changes. 

Cheers, 
Julien.