@thijssteel 2021-02-02
1. The code uses the zeros below the diagonal as scratch. In general, the amount of required scratch decreases with this PR, because the windows are smaller, but the minimal blocksize is now 4 instead of 3 (its actually possible to also make it 3, but 4 makes it a bit faster for the larger pencils). This blocksize occurs twice in the subdiagonal dimensions so we need to add 2 to ntiny in total. This is the same for the complex case because it also uses double shifts.

2. Yes, i'll change it.