@thijssteel 2021-02-03
oh, and to be clear, this isn't finished yet. I still need to investigate `NSMAX` and check other possible impacts.