@thijssteel 2021-02-04
Small note on NTINY and NSMAX in case anyone want to check my math.

As described in lawn 187, we need a block below the diagonal that allows for a KDU x KDU matrix for U, a KDU (or larger) x KDU matrix for W and some space below the diagonal so that it doesn't interfere with the bulge chasing.

For the old implementation we had the following constraint: N > KDU + 1 (For the hessenberg part on top) + 2 (For the bulges) + 2*KDU (For U and W). This leads to NS <= (N+6)/9. Note that this requires NTINY = 12, not 11, a minor bug in XLAQR0

In the new implementation, KDU = 2NS instead of 3NS - 3, this leads to NS <= (N-3)/6. This requires NTINY = 15, my previous calculation was wrong (didn't account for the error in the previous implementation and also missed the extra space required by the now larger Hessenberg part).

This PR should be ready for review now.